Fixed headers spacing and em dashes
According to this: https://en.oxforddictionaries.com/grammar/dash-and-em-dash and multiple other sites, em dashes should be used without whitespaces on either side. But in order for this to render as intended, there can't be a line break after the em dash, which @davste0816 has already noticed. I decided to remove all the line breaks after the em dashes. TBH, this still doesn't look quite right—it seems that the space between the em dashes and the letters is a little too small. I guess this is caused by the font?