CSS: change structure of the new box classes
In the markdown now, the class list will have to be specified with a space
in-between the two classes (i.e. "warning box") in order for the new condensed
styling to work.

The `display: inline-flex` was unnecessary for the ::before section and the
`background: 0 0` wasn't actually doing anything, so both were removed.